# apps/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.AssignMenusRequest:
    properties:
      menuIds:
        items:
          type: integer
        type: array
    required:
    - menuIds
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.CreateDeptRequest:
    properties:
      leader:
        maxLength: 64
        type: string
      name:
        maxLength: 64
        type: string
      parentId:
        type: integer
      phone:
        maxLength: 32
        type: string
      sort:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
    required:
    - name
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.CreateMenuRequest:
    properties:
      component:
        maxLength: 255
        type: string
      icon:
        maxLength: 100
        type: string
      isFrame:
        enum:
        - 0
        - 1
        type: integer
      name:
        maxLength: 64
        type: string
      parentId:
        type: integer
      path:
        maxLength: 200
        type: string
      perms:
        maxLength: 100
        type: string
      sort:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
      type:
        enum:
        - 1
        - 2
        - 3
        type: integer
      visible:
        enum:
        - 0
        - 1
        type: integer
    required:
    - name
    - type
    type: object
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.CreateRoleRequest:
    properties:
      code:
        description: Code 为权限判定依据，只允许字母/数字/下划线/冒号，避免与 Casbin 策略解析冲突。
        maxLength: 64
        type: string
      dataScope:
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        type: integer
      deptIds:
        items:
          type: integer
        type: array
      menuIds:
        items:
          type: integer
        type: array
      name:
        maxLength: 64
        type: string
      remark:
        maxLength: 255
        type: string
      sort:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
    required:
    - code
    - name
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.DataScopeRequest:
    properties:
      dataScope:
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        type: integer
      deptIds:
        description: DeptIDs 仅在 DataScope=2（自定义）时生效。
        items:
          type: integer
        type: array
    required:
    - dataScope
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.UpdateDeptRequest:
    properties:
      leader:
        maxLength: 64
        type: string
      name:
        maxLength: 64
        type: string
      parentId:
        type: integer
      phone:
        maxLength: 32
        type: string
      sort:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.UpdateMenuRequest:
    properties:
      component:
        maxLength: 255
        type: string
      icon:
        maxLength: 100
        type: string
      isFrame:
        enum:
        - 0
        - 1
        type: integer
      name:
        maxLength: 64
        type: string
      parentId:
        type: integer
      path:
        maxLength: 200
        type: string
      perms:
        maxLength: 100
        type: string
      sort:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
      type:
        enum:
        - 1
        - 2
        - 3
        type: integer
      visible:
        enum:
        - 0
        - 1
        type: integer
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.UpdateRoleRequest:
    properties:
      dataScope:
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        type: integer
      menuIds:
        items:
          type: integer
        type: array
      name:
        maxLength: 64
        type: string
      remark:
        maxLength: 255
        type: string
      sort:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_vo.RoleDetail:
    properties:
      code:
        type: string
      createdAt:
        type: string
      dataScope:
        type: integer
      deptIds:
        description: DeptIDs 仅在 DataScope 为自定义时有意义。
        items:
          type: integer
        type: array
      id:
        type: integer
      menuIds:
        items:
          type: integer
        type: array
      name:
        type: string
      remark:
        type: string
      sort:
        type: integer
      status:
        type: integer
    type: object
  github_com_fangzhengcn_workbackend_apps_backend_internal_vo.RoleItem:
    properties:
      code:
        type: string
      createdAt:
        type: string
      dataScope:
        type: integer
      id:
        type: integer
      name:
        type: string
      remark:
        type: string
      sort:
        type: integer
      status:
        type: integer
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  github_com_fangzhengcn_workbackend_apps_backend_pkg_response.PageData:
    properties:
      list: {}
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
info:
  contact: {}
  description: 基于 RBAC 的通用后台管理系统接口
  title: 权限管理后台 API
  version: "1.0"
paths:
...
  /depts:
    post:
      consumes:
      - application/json
      parameters:
      - description: 部门信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.CreateDeptRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 新增部门
      tags:
      - 部门管理
  /depts/{id}:
    delete:
      parameters:
      - description: 部门ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 删除部门
      tags:
      - 部门管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 部门ID
        in: path
        name: id
        required: true
        type: integer
      - description: 部门信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.UpdateDeptRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 修改部门
      tags:
      - 部门管理
...
  /menus:
    post:
      consumes:
      - application/json
      parameters:
      - description: 菜单信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.CreateMenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 新增菜单
      tags:
      - 菜单管理
  /menus/{id}:
    delete:
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 删除菜单
      tags:
      - 菜单管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 菜单信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.UpdateMenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 修改菜单
      tags:
      - 菜单管理
...
  /roles:
    get:
      parameters:
      - in: query
        maxLength: 64
        name: code
        type: string
      - in: query
        maxLength: 64
        name: name
        type: string
      - in: query
        minimum: 1
        name: page
        type: integer
      - in: query
        maximum: 200
        minimum: 1
        name: size
        type: integer
      - enum:
        - 0
        - 1
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_vo.RoleItem'
                        type: array
                    type: object
              type: object
      summary: 角色列表
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      parameters:
      - description: 角色信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 新增角色
      tags:
      - 角色管理
  /roles/{id}:
    delete:
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 删除角色
      tags:
      - 角色管理
    get:
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_vo.RoleDetail'
              type: object
      summary: 角色详情
      tags:
      - 角色管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 修改角色
      tags:
      - 角色管理
  /roles/{id}/data-scope:
    put:
      consumes:
      - application/json
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 数据范围
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.DataScopeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 设置数据权限
      tags:
      - 角色管理
  /roles/{id}/menus:
    get:
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
            - properties:
                data:
                  items:
                    type: integer
                  type: array
              type: object
      summary: 角色菜单ID
      tags:
      - 角色管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 菜单ID集合
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.AssignMenusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 分配菜单权限
      tags:
      - 角色管理
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"